docs: create SETTINGS.md to describe the default repo settings
Create a new document, SETTINGS.md, that describes the settings that should be configured in each repo. This will serve as a starting point for creating a set of consistent settings across all repos in the org. These settings can be further tweaked as needed to fit the needs of each repo.